refactor: move crafting UI logic to component
